Rename knex.ts to knexFile.ts
